requirements-simple.txt
# Core Web Framework
fastapi
uvicorn[standard]
pydantic

# Document Processing
pymupdf
python-pptx
Pillow

# UI
streamlit
requests

# LLaMA Index & RAG
llama-index-core
llama-index-readers-file
llama-index-llms-huggingface

# ML/DL
torch
transformers
numpy
scikit-learn

# Utilities
python-dotenv
